WIP: Break up: Separate (de)serialisation
